pyproject.toml
[tool.poetry]
name = "myapp"
version = "0.1.0"
description = ""
authors = ["r_nakamaru <r.nakamaru@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.19.0"
pyrebase4 = "^4.6.0"
google-cloud-bigquery = "^3.6.0"
pandas = "^1.5.3"
db-dtypes = "^1.0.5"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.1.0"
flake8 = "^6.0.0"
pip-audit = "^2.4.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
